Finish incorporating new Tags framework
dcs init cleanup and cleanliness changes
